// src/types/postTypes.ts
export interface PostItemType {
  id: string;
  userId: string;
  username: string;
  avatar: string;
  shopName: string;
  shopId: string;
  images: string[];
  comments: number;
  date: string;
  text: string;
  likes: number;
  didLike: boolean;
  premium: boolean;
}

export interface PostDataResponse {
  hasMore: boolean;
  data: PostItemType[];
}

export interface PostItemPops {
  post: PostItemType;
}